Move BaseCharAtlas to browser
