+++ src/renderer/atlas/Types.d.ts
@@ -6,16 +6,6 @@
 import { FontWeight } from 'xterm';
 import { IColorSet } from 'browser/Types';
 
-export interface IGlyphIdentifier {
-  chars: string;
-  code: number;
-  bg: number;
-  fg: number;
-  bold: boolean;
-  dim: boolean;
-  italic: boolean;
-}
-
 export interface ICharAtlasConfig {
   devicePixelRatio: number;
   fontSize: number;
